1.实现发布商品（涉及SPU SKU相关表格的增删改查） 2.远程调用coupon模块的积分、优惠、满减信息管理 3.库存管理下的商品库存与采购单维护 4.采购单的合并、领取、完成功能 5.商品维护下SPU管理实现规格的查看和修改

+++ gulimall-product/src/main/java/com/gulimall/product/domain/SkuImages.java
@@ -39,6 +39,6 @@ private static final long serialVersionUID = 1L;
 
     /** 默认图[0 - 不是默认图，1 - 是默认图] */
     @Excel(name = "默认图[0 - 不是默认图，1 - 是默认图]")
-    private Long defaultImg;
+    private Integer defaultImg;
 
 }
